format codes

+++ contracts/Compound/CERC20Interface.sol
@@ -3,9 +3,12 @@ pragma solidity ^0.8.14;
 
 import "./CTokenInterface.sol";
 
-interface CERC20Interface is CTokenInterface{    
-    function mint(uint mintAmount) external returns (uint);
-    function redeem(uint redeemTokens) external returns (uint);
-    function borrow(uint borrowAmount) external returns (uint);
-    function repayBorrow(uint repayAmount) external returns (uint);
+interface CERC20Interface is CTokenInterface {
+    function mint(uint256 mintAmount) external returns (uint256);
+
+    function redeem(uint256 redeemTokens) external returns (uint256);
+
+    function borrow(uint256 borrowAmount) external returns (uint256);
+
+    function repayBorrow(uint256 repayAmount) external returns (uint256);
 }
