+++ contracts/Compound/CETHInterface.sol
@@ -3,9 +3,12 @@ pragma solidity ^0.8.14;
 
 import "./CTokenInterface.sol";
 
-interface CETHInterface is CTokenInterface{    
-    function mint() external payable returns (uint);
-    function redeem(uint redeemTokens) external returns (uint);
-    function borrow(uint borrowAmount) external returns (uint);
-    function repayBorrow() external payable returns (uint);
+interface CETHInterface is CTokenInterface {
+    function mint() external payable returns (uint256);
+
+    function redeem(uint256 redeemTokens) external returns (uint256);
+
+    function borrow(uint256 borrowAmount) external returns (uint256);
+
+    function repayBorrow() external payable returns (uint256);
 }
