+++ contracts/Compound/IOracle.sol
@@ -4,7 +4,8 @@ pragma solidity ^0.8.14;
 import "./CTokenInterface.sol";
 
 interface IOracle {
-    function getUnderlyingPrice(CTokenInterface cToken) external view returns (uint);
+    function getUnderlyingPrice(CTokenInterface cToken)
+        external
+        view
+        returns (uint256);
 }
-
-
