+++ contracts/libraries/math.sol
@@ -1,16 +1,16 @@
 // SPDX-License-Identifier: UNLICENSED
-pragma solidity >= 0.8.0;
+pragma solidity >=0.8.0;
 
-library Math{
-    function divCeil(uint x, uint y) internal pure returns (uint){
+library Math {
+    function divCeil(uint256 x, uint256 y) internal pure returns (uint256) {
         return (x + y - 1) / y;
     }
 
     // babylonian method (https://en.wikipedia.org/wiki/Methods_of_computing_square_roots#Babylonian_method)
-    function sqrt(uint y) internal pure returns (uint z) {
+    function sqrt(uint256 y) internal pure returns (uint256 z) {
         if (y > 3) {
             z = y;
-            uint x = y / 2 + 1;
+            uint256 x = y / 2 + 1;
             while (x < z) {
                 z = x;
                 x = (y / x + x) / 2;
